Everything working!!
Working with Gemini. The only planned change at this point is that I want to take out all the console messages, which were great for debugging but not great for production as it gives the user way too much info about what's going on behind the scenes if they hit F12.